+++ demo/js/gui.js
@@ -1,11 +1,11 @@
 // Generated by CoffeeScript 1.9.2
-(function() {
+(function () {
   var isInArray, konamiCode;
 
-  window.loadGUI = function() {
+  window.loadGUI = function () {
     var f, ff, fm, fmm, gui_f, initPage, resizeGUI;
     window.p = {};
-    p.exportConfig = function(type) {
+    p.exportConfig = function (type) {
       var data, data_css, pJS_GUI_Export, pJS_GUI_tmp, ratio;
       pJS_GUI_Export = {};
       pJS_GUI_tmp = window.pJS_GUI;
@@ -40,11 +40,11 @@
         return data;
       }
     };
-    p.exportConfigJSON = function() {
+    p.exportConfigJSON = function () {
       return p.exportConfig('json');
     };
-    p.update = function() {
-      return pJS_GUI.particlesRefresh();
+    p.update = function () {
+      return pJS_GUI.fn.particles.refresh();
     };
     window.gui = new dat.GUI({
       autoPlace: true,
@@ -91,49 +91,49 @@
       repulse: f.interactivity.modes.addFolder('repulse'),
       push: f.interactivity.modes.addFolder('push'),
       remove: f.interactivity.modes.addFolder('remove'),
-	  parallax: f.interactivity.events_onhover.addFolder('parallax')
+      parallax: f.interactivity.events_onhover.addFolder('parallax')
     };
     fmm = {
       grab_linelinked: fm.grab.addFolder('line_linked')
     };
-    gui.add(pJS_GUI, 'retina_detect').name('retina_detect').onChange(function(value) {
+    gui.add(pJS_GUI, 'retina_detect').name('retina_detect').onChange(function (value) {
       console.log(pJS_GUI.retina_detect);
       pJS_GUI.retina_detect = value;
       return p.update();
     });
-    f.particles.number.add(pJS_GUI.particles.number, 'value', 0, 600).name('value').step(1).onChange(function(value) {
+    f.particles.number.add(pJS_GUI.particles.number, 'value', 0, 600).name('value').step(1).onChange(function (value) {
       pJS_GUI.particles.number.value = value;
       return p.update();
     });
-    f.particles.number.add(pJS_GUI.particles.number.density, 'enable').name('density.enable').onChange(function(value) {
+    f.particles.number.add(pJS_GUI.particles.number.density, 'enable').name('density.enable').onChange(function (value) {
       pJS_GUI.particles.number.density.enable = value;
       return p.update();
     });
-    f.particles.number.add(pJS_GUI.particles.number.density, 'value_area', 0, 10000).name('density.value_area').onChange(function(value) {
+    f.particles.number.add(pJS_GUI.particles.number.density, 'value_area', 0, 10000).name('density.value_area').onChange(function (value) {
       pJS_GUI.particles.number.density.value_area = value;
       return p.update();
     });
-    f.particles.color.addColor(pJS_GUI.particles.color, 'value').name('value (single string)').onChange(function(value) {
+    f.particles.color.addColor(pJS_GUI.particles.color, 'value').name('value (single string)').onChange(function (value) {
       pJS_GUI.particles.color.value = value;
       return p.update();
     });
-    f.particles.shape.add(pJS_GUI.particles.shape, 'type', ['circle', 'edge', 'triangle', 'polygon', 'star', 'image']).name('type').onChange(function(value) {
+    f.particles.shape.add(pJS_GUI.particles.shape, 'type', ['circle', 'edge', 'triangle', 'polygon', 'star', 'image']).name('type').onChange(function (value) {
       pJS_GUI.particles.shape.type = value;
       return p.update();
     });
-    ff.shape_polygon.add(pJS_GUI.particles.shape.polygon, 'nb_sides', 3, 12).step(1).name('polygon.nb_sides').onChange(function(value) {
+    ff.shape_polygon.add(pJS_GUI.particles.shape.polygon, 'nb_sides', 3, 12).step(1).name('polygon.nb_sides').onChange(function (value) {
       pJS_GUI.particles.shape.polygon.nb_sides = value;
       return p.update();
     });
-    ff.shape_stroke.add(pJS_GUI.particles.shape.stroke, 'width', 0, 20).step(1).name('stroke.width').onChange(function(value) {
+    ff.shape_stroke.add(pJS_GUI.particles.shape.stroke, 'width', 0, 20).step(1).name('stroke.width').onChange(function (value) {
       pJS_GUI.particles.shape.stroke.width = value;
       return p.update();
     });
-    ff.shape_stroke.addColor(pJS_GUI.particles.shape.stroke, 'color').name('stroke.color').onChange(function(value) {
+    ff.shape_stroke.addColor(pJS_GUI.particles.shape.stroke, 'color').name('stroke.color').onChange(function (value) {
       pJS_GUI.particles.shape.stroke.color = value;
       return p.update();
     });
-    ff.shape_image.add(pJS_GUI.particles.shape.image, 'src').name('image.src').onChange(function(value) {
+    ff.shape_image.add(pJS_GUI.particles.shape.image, 'src').name('image.src').onChange(function (value) {
       var type;
       pJS_GUI.particles.shape.image.src = value;
       if (pJS_GUI.particles.shape.type === 'image') {
@@ -144,194 +144,194 @@
       }
       return p.update();
     });
-    ff.shape_image.add(pJS_GUI.particles.shape.image, 'width').name('image.width').onChange(function(value) {
+    ff.shape_image.add(pJS_GUI.particles.shape.image, 'width').name('image.width').onChange(function (value) {
       pJS_GUI.particles.shape.image.width = value;
       return p.update();
     });
-    ff.shape_image.add(pJS_GUI.particles.shape.image, 'height').name('image.height').onChange(function(value) {
+    ff.shape_image.add(pJS_GUI.particles.shape.image, 'height').name('image.height').onChange(function (value) {
       pJS_GUI.particles.shape.image.height = value;
       return p.update();
     });
-    f.particles.size.add(pJS_GUI.tmp.obj, 'size_value', 0, 500).name('value').onChange(function(value) {
+    f.particles.size.add(pJS_GUI.tmp.obj, 'size_value', 0, 500).name('value').onChange(function (value) {
       pJS_GUI.tmp.obj.size_value = value;
       return p.update();
     });
-    f.particles.size.add(pJS_GUI.particles.size, 'random').name('random').onChange(function(value) {
+    f.particles.size.add(pJS_GUI.particles.size, 'random').name('random').onChange(function (value) {
       pJS_GUI.particles.size.random = value;
       return p.update();
     });
-    ff.size_anim.add(pJS_GUI.particles.size.anim, 'enable').name('anim.enable').onChange(function(value) {
+    ff.size_anim.add(pJS_GUI.particles.size.anim, 'enable').name('anim.enable').onChange(function (value) {
       pJS_GUI.particles.size.anim.enable = value;
       return p.update();
     });
-    ff.size_anim.add(pJS_GUI.tmp.obj, 'size_anim_speed', 0, 300).name('anim.speed').onChange(function(value) {
+    ff.size_anim.add(pJS_GUI.tmp.obj, 'size_anim_speed', 0, 300).name('anim.speed').onChange(function (value) {
       pJS_GUI.tmp.obj.size_anim_speed = value;
       return p.update();
     });
-    ff.size_anim.add(pJS_GUI.particles.size.anim, 'size_min', 0, 100).name('anim.size_min').onChange(function(value) {
+    ff.size_anim.add(pJS_GUI.particles.size.anim, 'size_min', 0, 100).name('anim.size_min').onChange(function (value) {
       pJS_GUI.particles.size.anim.size_min = value;
       return p.update();
     });
-    ff.size_anim.add(pJS_GUI.particles.size.anim, 'sync').name('anim.sync').onChange(function(value) {
+    ff.size_anim.add(pJS_GUI.particles.size.anim, 'sync').name('anim.sync').onChange(function (value) {
       pJS_GUI.particles.size.anim.sync = value;
       return p.update();
     });
-    f.particles.opacity.add(pJS_GUI.particles.opacity, 'value', 0, 1).name('value').onChange(function(value) {
+    f.particles.opacity.add(pJS_GUI.particles.opacity, 'value', 0, 1).name('value').onChange(function (value) {
       pJS_GUI.particles.opacity.value = value;
       return p.update();
     });
-    f.particles.opacity.add(pJS_GUI.particles.opacity, 'random').name('random').onChange(function(value) {
+    f.particles.opacity.add(pJS_GUI.particles.opacity, 'random').name('random').onChange(function (value) {
       pJS_GUI.particles.opacity.random = value;
       return p.update();
     });
-    ff.opacity_anim.add(pJS_GUI.particles.opacity.anim, 'enable').name('anim.enable').onChange(function(value) {
+    ff.opacity_anim.add(pJS_GUI.particles.opacity.anim, 'enable').name('anim.enable').onChange(function (value) {
       pJS_GUI.particles.opacity.anim.enable = value;
       return p.update();
     });
-    ff.opacity_anim.add(pJS_GUI.particles.opacity.anim, 'speed', 0, 10).name('anim.speed').onChange(function(value) {
+    ff.opacity_anim.add(pJS_GUI.particles.opacity.anim, 'speed', 0, 10).name('anim.speed').onChange(function (value) {
       pJS_GUI.particles.opacity.anim.speed = value;
       return p.update();
     });
-    ff.opacity_anim.add(pJS_GUI.particles.opacity.anim, 'opacity_min', 0, 1).name('anim.opacity_min').onChange(function(value) {
+    ff.opacity_anim.add(pJS_GUI.particles.opacity.anim, 'opacity_min', 0, 1).name('anim.opacity_min').onChange(function (value) {
       pJS_GUI.particles.opacity.anim.opacity_min = value;
       return p.update();
     });
-    ff.opacity_anim.add(pJS_GUI.particles.opacity.anim, 'sync').name('anim.sync').onChange(function(value) {
+    ff.opacity_anim.add(pJS_GUI.particles.opacity.anim, 'sync').name('anim.sync').onChange(function (value) {
       pJS_GUI.particles.opacity.anim.sync = value;
       return p.update();
     });
-    f.particles.line_linked.add(pJS_GUI.particles.line_linked, 'enable').name('enable_auto').onChange(function(value) {
+    f.particles.line_linked.add(pJS_GUI.particles.line_linked, 'enable').name('enable_auto').onChange(function (value) {
       pJS_GUI.particles.line_linked.enable = value;
       return p.update();
     });
-    f.particles.line_linked.add(pJS_GUI.tmp.obj, 'line_linked_distance', 0, 2000).name('distance').onChange(function(value) {
+    f.particles.line_linked.add(pJS_GUI.tmp.obj, 'line_linked_distance', 0, 2000).name('distance').onChange(function (value) {
       pJS_GUI.tmp.obj.line_linked_distance = value;
       return p.update();
     });
-    f.particles.line_linked.addColor(pJS_GUI.particles.line_linked, 'color').name('color').onChange(function(value) {
+    f.particles.line_linked.addColor(pJS_GUI.particles.line_linked, 'color').name('color').onChange(function (value) {
       pJS_GUI.particles.line_linked.color = value;
       return p.update();
     });
-    f.particles.line_linked.add(pJS_GUI.particles.line_linked, 'opacity', 0, 1).name('opacity').onChange(function(value) {
+    f.particles.line_linked.add(pJS_GUI.particles.line_linked, 'opacity', 0, 1).name('opacity').onChange(function (value) {
       pJS_GUI.particles.line_linked.opacity = value;
       return p.update();
     });
-    f.particles.line_linked.add(pJS_GUI.tmp.obj, 'line_linked_width', 0, 20).name('width').onChange(function(value) {
+    f.particles.line_linked.add(pJS_GUI.tmp.obj, 'line_linked_width', 0, 20).name('width').onChange(function (value) {
       pJS_GUI.tmp.obj.line_linked_width = value;
       return p.update();
     });
-    f.particles.move.add(pJS_GUI.particles.move, 'enable').name('enable').onChange(function(value) {
+    f.particles.move.add(pJS_GUI.particles.move, 'enable').name('enable').onChange(function (value) {
       pJS_GUI.particles.move.enable = value;
       return p.update();
     });
-    f.particles.move.add(pJS_GUI.particles.move, 'direction', ['none', 'top', 'top-right', 'right', 'bottom-right', 'bottom', 'bottom-left', 'left', 'top-left']).name('direction').onChange(function(value) {
+    f.particles.move.add(pJS_GUI.particles.move, 'direction', ['none', 'top', 'top-right', 'right', 'bottom-right', 'bottom', 'bottom-left', 'left', 'top-left']).name('direction').onChange(function (value) {
       pJS_GUI.particles.move.direction = value;
       return p.update();
     });
-    f.particles.move.add(pJS_GUI.particles.move, 'random').name('random').onChange(function(value) {
+    f.particles.move.add(pJS_GUI.particles.move, 'random').name('random').onChange(function (value) {
       pJS_GUI.particles.move.random = value;
       return p.update();
     });
-    f.particles.move.add(pJS_GUI.particles.move, 'straight').name('straight').onChange(function(value) {
+    f.particles.move.add(pJS_GUI.particles.move, 'straight').name('straight').onChange(function (value) {
       pJS_GUI.particles.move.straight = value;
       return p.update();
     });
-    f.particles.move.add(pJS_GUI.tmp.obj, 'move_speed', 0, 200).name('speed').onChange(function(value) {
+    f.particles.move.add(pJS_GUI.tmp.obj, 'move_speed', 0, 200).name('speed').onChange(function (value) {
       pJS_GUI.tmp.obj.move_speed = value;
       return p.update();
     });
-    f.particles.move.add(pJS_GUI.particles.move, 'out_mode', ['out', 'bounce']).name('out_mode').onChange(function(value) {
+    f.particles.move.add(pJS_GUI.particles.move, 'out_mode', ['out', 'bounce']).name('out_mode').onChange(function (value) {
       pJS_GUI.particles.move.out_mode = value;
       return p.update();
     });
-    f.particles.move.add(pJS_GUI.particles.move.attract, 'enable').name('attract.enable').onChange(function(value) {
+    f.particles.move.add(pJS_GUI.particles.move.attract, 'enable').name('attract.enable').onChange(function (value) {
       pJS_GUI.particles.move.attract.enable = value;
       return p.update();
     });
-    f.particles.move.add(pJS_GUI.particles.move.attract, 'rotateX', 0, 10000).name('attract.rotateX').onChange(function(value) {
+    f.particles.move.add(pJS_GUI.particles.move.attract, 'rotateX', 0, 10000).name('attract.rotateX').onChange(function (value) {
       pJS_GUI.particles.move.attract.rotateX = value;
       return p.update();
     });
-    f.particles.move.add(pJS_GUI.particles.move.attract, 'rotateY', 0, 10000).name('attract.rotateY').onChange(function(value) {
+    f.particles.move.add(pJS_GUI.particles.move.attract, 'rotateY', 0, 10000).name('attract.rotateY').onChange(function (value) {
       pJS_GUI.particles.move.attract.rotateY = value;
       return p.update();
     });
-    gui_f.interactivity.add(pJS_GUI.interactivity, 'detect_on', ['window', 'canvas']).name('detect_on').onChange(function(value) {
+    gui_f.interactivity.add(pJS_GUI.interactivity, 'detect_on', ['window', 'canvas']).name('detect_on').onChange(function (value) {
       pJS_GUI.interactivity.detect_on = value;
       return p.update();
     });
-    f.interactivity.events_onhover.add(pJS_GUI.interactivity.events.onhover, 'enable').name('enable').onChange(function(value) {
+    f.interactivity.events_onhover.add(pJS_GUI.interactivity.events.onhover, 'enable').name('enable').onChange(function (value) {
       pJS_GUI.interactivity.events.onhover.enable = value;
       return p.update();
     });
-    f.interactivity.events_onhover.add(pJS_GUI.interactivity.events.onhover, 'mode', ['grab', 'bubble', 'repulse']).name('mode').onChange(function(value) {
+    f.interactivity.events_onhover.add(pJS_GUI.interactivity.events.onhover, 'mode', ['grab', 'bubble', 'repulse']).name('mode').onChange(function (value) {
       pJS_GUI.interactivity.events.onhover.mode = value;
       return p.update();
     });
-	fm.parallax.add(pJS_GUI.interactivity.events.onhover.parallax, 'enable').name('enable').onChange(function(value) {
+    fm.parallax.add(pJS_GUI.interactivity.events.onhover.parallax, 'enable').name('enable').onChange(function (value) {
       pJS_GUI.interactivity.events.onhover.parallax.enable = value;
       return p.update();
     });
-	fm.parallax.add(pJS_GUI.interactivity.events.onhover.parallax, 'force', 0, 100).name('force').onChange(function(value) {
+    fm.parallax.add(pJS_GUI.interactivity.events.onhover.parallax, 'force', 0, 100).name('force').onChange(function (value) {
       pJS_GUI.interactivity.events.onhover.parallax.force = value;
       return p.update();
     });
-	fm.parallax.add(pJS_GUI.interactivity.events.onhover.parallax, 'smooth', 0, 100).name('smooth').onChange(function(value) {
+    fm.parallax.add(pJS_GUI.interactivity.events.onhover.parallax, 'smooth', 0, 100).name('smooth').onChange(function (value) {
       pJS_GUI.interactivity.events.onhover.parallax.smooth = value;
       return p.update();
     });
-    f.interactivity.events_onclick.add(pJS_GUI.interactivity.events.onclick, 'enable').name('enable').onChange(function(value) {
+    f.interactivity.events_onclick.add(pJS_GUI.interactivity.events.onclick, 'enable').name('enable').onChange(function (value) {
       pJS_GUI.interactivity.events.onclick.enable = value;
       return p.update();
     });
-    f.interactivity.events_onclick.add(pJS_GUI.interactivity.events.onclick, 'mode', ['push', 'remove', 'bubble', 'repulse']).name('mode').onChange(function(value) {
+    f.interactivity.events_onclick.add(pJS_GUI.interactivity.events.onclick, 'mode', ['push', 'remove', 'bubble', 'repulse']).name('mode').onChange(function (value) {
       pJS_GUI.interactivity.events.onclick.mode = value;
       return p.update();
     });
-    fm.grab.add(pJS_GUI.tmp.obj, 'mode_grab_distance', 0, 1500).name('distance').onChange(function(value) {
+    fm.grab.add(pJS_GUI.tmp.obj, 'mode_grab_distance', 0, 1500).name('distance').onChange(function (value) {
       pJS_GUI.tmp.obj.mode_grab_distance = value;
       return p.update();
     });
-    fmm.grab_linelinked.add(pJS_GUI.interactivity.modes.grab.line_linked, 'opacity', 0, 1).name('opacity').onChange(function(value) {
+    fmm.grab_linelinked.add(pJS_GUI.interactivity.modes.grab.line_linked, 'opacity', 0, 1).name('opacity').onChange(function (value) {
       pJS_GUI.interactivity.modes.grab.line_linked.opacity = value;
       return p.update();
     });
-    fm.bubble.add(pJS_GUI.tmp.obj, 'mode_bubble_distance', 0, 1500).name('distance').onChange(function(value) {
+    fm.bubble.add(pJS_GUI.tmp.obj, 'mode_bubble_distance', 0, 1500).name('distance').onChange(function (value) {
       pJS_GUI.tmp.obj.mode_bubble_distance = value;
       return p.update();
     });
-    fm.bubble.add(pJS_GUI.tmp.obj, 'mode_bubble_size', 0, 500).name('size').onChange(function(value) {
+    fm.bubble.add(pJS_GUI.tmp.obj, 'mode_bubble_size', 0, 500).name('size').onChange(function (value) {
       pJS_GUI.tmp.obj.mode_bubble_size = value;
       return p.update();
     });
-    fm.bubble.add(pJS_GUI.interactivity.modes.bubble, 'opacity', 0, 1).name('opacity').onChange(function(value) {
+    fm.bubble.add(pJS_GUI.interactivity.modes.bubble, 'opacity', 0, 1).name('opacity').onChange(function (value) {
       pJS_GUI.interactivity.modes.bubble.opacity = value;
       return p.update();
     });
-    fm.bubble.add(pJS_GUI.interactivity.modes.bubble, 'duration', 0, 10).name('duration (sec)').onChange(function(value) {
+    fm.bubble.add(pJS_GUI.interactivity.modes.bubble, 'duration', 0, 10).name('duration (sec)').onChange(function (value) {
       pJS_GUI.interactivity.modes.bubble.duration = value;
       return p.update();
     });
-    fm.repulse.add(pJS_GUI.tmp.obj, 'mode_repulse_distance', 0, 1000).name('distance').onChange(function(value) {
+    fm.repulse.add(pJS_GUI.tmp.obj, 'mode_repulse_distance', 0, 1000).name('distance').onChange(function (value) {
       pJS_GUI.tmp.obj.mode_repulse_distance = value;
       return p.update();
     });
-    gui_f.pagecss.addColor(pJS_GUI.config_demo, 'background_color').name('background-color').onChange(function(value) {
+    gui_f.pagecss.addColor(pJS_GUI.config_demo, 'background_color').name('background-color').onChange(function (value) {
       return $('#particles-js').css('background-color', value);
     });
-    gui_f.pagecss.add(pJS_GUI.config_demo, 'background_image').name('background-image url').onChange(function(value) {
+    gui_f.pagecss.add(pJS_GUI.config_demo, 'background_image').name('background-image url').onChange(function (value) {
       return $('#particles-js').css('background-image', 'url(' + value + ')');
     });
-    gui_f.pagecss.add(pJS_GUI.config_demo, 'background_size').name('background-size').onChange(function(value) {
+    gui_f.pagecss.add(pJS_GUI.config_demo, 'background_size').name('background-size').onChange(function (value) {
       return $('#particles-js').css('background-size', value);
     });
-    gui_f.pagecss.add(pJS_GUI.config_demo, 'background_position').name('background-position').onChange(function(value) {
+    gui_f.pagecss.add(pJS_GUI.config_demo, 'background_position').name('background-position').onChange(function (value) {
       return $('#particles-js').css('background-position', value);
     });
-    gui_f.pagecss.add(pJS_GUI.config_demo, 'background_repeat').name('background-repeat').onChange(function(value) {
+    gui_f.pagecss.add(pJS_GUI.config_demo, 'background_repeat').name('background-repeat').onChange(function (value) {
       return $('#particles-js').css('background-repeat', value);
     });
-    gui_f.pagecss.add(pJS_GUI.config_demo, 'hide_card').name('hide card').onChange(function(value) {
+    gui_f.pagecss.add(pJS_GUI.config_demo, 'hide_card').name('hide card').onChange(function (value) {
       if (value) {
         $('.panel').fadeOut(200);
         return $('.js-box-bottom').fadeIn(200);
@@ -342,7 +342,7 @@
     });
     gui.add(pJS_GUI.fn.vendors, 'exportImg').name('Export image (png)');
     gui.add(p, 'exportConfigJSON').name('Download current config (json)');
-    initPage = function() {
+    initPage = function () {
       var box_bottom, config, page, panel;
       page = $(pJS_GUI.canvas.el).parent();
       panel = $('.panel');
@@ -366,19 +366,19 @@
       return $('body').addClass('loaded');
     };
     initPage();
-    resizeGUI = function() {
+    resizeGUI = function () {
       return $('.dg.main > ul').css('max-height', window.innerHeight - 105);
     };
     resizeGUI();
-    $(window).on('resize', function() {
+    $(window).on('resize', function () {
       return resizeGUI();
     });
-    return $('.close-button').on('click', function() {
+    return $('.close-button').on('click', function () {
       return $('.box-top').css('width', $('.dg.main').width());
     });
   };
 
-  window.loadStats = function() {
+  window.loadStats = function () {
     var count_particles, stats, update;
     stats = new Stats;
     stats.setMode(0);
@@ -387,7 +387,7 @@
     stats.domElement.style.top = '0px';
     document.body.appendChild(stats.domElement);
     count_particles = $('.js-count-particles');
-    update = function() {
+    update = function () {
       stats.begin();
       stats.end();
       if (window.pJS_GUI.particles && window.pJS_GUI.particles.array) {
@@ -398,8 +398,8 @@
     return requestAnimationFrame(update);
   };
 
-  (function(console) {
-    console.save = function(data, filename) {
+  (function (console) {
+    console.save = function (data, filename) {
       var a, blob, e;
       if (!data) {
         console.error('Console.save: No data');
@@ -424,11 +424,11 @@
     };
   })(console);
 
-  isInArray = function(value, array) {
+  isInArray = function (value, array) {
     return array.indexOf(value) > -1;
   };
 
-  $('.js-dp-label').on('click', function() {
+  $('.js-dp-label').on('click', function () {
     var dropdown, id, t;
     t = $(this);
     id = $(this).attr('data-dp');
@@ -443,13 +443,13 @@
     return t.toggleClass('on');
   });
 
-  konamiCode = function() {
+  konamiCode = function () {
     var k, n, pjs, pjs_style, playAudio, start, stop;
     k = [38, 38, 40, 40, 37, 39, 37, 39, 66, 65];
     n = 0;
     pjs = $('#particles-js');
     pjs_style = null;
-    playAudio = function() {
+    playAudio = function () {
       var audio, src_path;
       audio = document.createElement('audio');
       src_path = window.location.origin + window.location.pathname + 'assets/medias/nyancat-loop.mp3';
@@ -458,7 +458,7 @@
       audio.addEventListener('ended', stop);
       return audio.play();
     };
-    start = function(audio) {
+    start = function (audio) {
       window.konami = true;
       pjs_style = pjs.attr('style');
       pjs.addClass('troll').css({
@@ -469,14 +469,14 @@
       });
       $('.panel').fadeOut(200);
       $('.js-box-bottom').fadeIn(200);
-      return $(document).on('click keydown', function() {
+      return $(document).on('click keydown', function () {
         if (window.konami) {
           audio.pause();
           return stop();
         }
       });
     };
-    stop = function() {
+    stop = function () {
       window.konami = false;
       pjs.removeClass('troll').attr('style', pjs_style);
       if (!pJS_GUI.config_demo.hide_card) {
@@ -484,7 +484,7 @@
         return $('.js-box_bottom').fadeOut(200);
       }
     };
-    return $(document).on('keydown', function(e) {
+    return $(document).on('keydown', function (e) {
       if (e.keyCode === k[n++]) {
         if (n === k.length) {
           playAudio();
@@ -497,8 +497,8 @@
     });
   };
 
-  particlesJS.load('particles-js', 'particles.json', function() {
-    return $(function() {
+  particlesJS.load('particles-js', 'particles.json', function () {
+    return $(function () {
       window.loadGUI();
       window.loadStats();
       return konamiCode();
