+++ src/particles.js
@@ -8,14 +8,14 @@
 /* ----------------------------------------------- */
 'use strict';
 
-import pJS from "./pjs";
+import { pJS } from "./pjs";
 
 /* ---------- global functions - vendors ------------ */
 
-Object.deepExtend = function(destination, source) {
+Object.deepExtend = function (destination, source) {
   for (var property in source) {
     if (source[property] && source[property].constructor &&
-     source[property].constructor === Object) {
+      source[property].constructor === Object) {
       destination[property] = destination[property] || {};
       this.deepExtend(destination[property], source[property]);
     } else {
@@ -25,38 +25,38 @@ Object.deepExtend = function(destination, source) {
   return destination;
 };
 
-window.requestAnimFrame = (function(){
-  return  window.requestAnimationFrame ||
+window.requestAnimFrame = (function () {
+  return window.requestAnimationFrame ||
     window.webkitRequestAnimationFrame ||
-    window.mozRequestAnimationFrame    ||
-    window.oRequestAnimationFrame      ||
-    window.msRequestAnimationFrame     ||
-    function(callback){
+    window.mozRequestAnimationFrame ||
+    window.oRequestAnimationFrame ||
+    window.msRequestAnimationFrame ||
+    function (callback) {
       window.setTimeout(callback, 1000 / 60);
     };
 })();
 
-window.cancelRequestAnimFrame = ( function() {
-  return window.cancelAnimationFrame         ||
+window.cancelRequestAnimFrame = (function () {
+  return window.cancelAnimationFrame ||
     window.webkitCancelRequestAnimationFrame ||
-    window.mozCancelRequestAnimationFrame    ||
-    window.oCancelRequestAnimationFrame      ||
-    window.msCancelRequestAnimationFrame     ||
+    window.mozCancelRequestAnimationFrame ||
+    window.oCancelRequestAnimationFrame ||
+    window.msCancelRequestAnimationFrame ||
     clearTimeout
-} )();
+})();
 
-function hexToRgb(hex){
+function hexToRgb(hex) {
   // By Tim Down - http://stackoverflow.com/a/5624139/3493650
   // Expand shorthand form (e.g. "03F") to full form (e.g. "0033FF")
   var shorthandRegex = /^#?([a-f\d])([a-f\d])([a-f\d])$/i;
-  hex = hex.replace(shorthandRegex, function(m, r, g, b) {
-     return r + r + g + g + b + b;
+  hex = hex.replace(shorthandRegex, function (m, r, g, b) {
+    return r + r + g + g + b + b;
   });
   var result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
   return result ? {
-      r: parseInt(result[1], 16),
-      g: parseInt(result[2], 16),
-      b: parseInt(result[3], 16)
+    r: parseInt(result[1], 16),
+    g: parseInt(result[2], 16),
+    b: parseInt(result[3], 16)
   } : null;
 };
 
@@ -73,29 +73,29 @@ function isInArray(value, array) {
 
 window.pJSDom = [];
 
-window.particlesJS = function(tag_id, params){
+window.particlesJS = function (tag_id, params) {
 
   //console.log(params);
 
   /* no string id? so it's object params, and set the id with default id */
-  if(typeof(tag_id) != 'string'){
+  if (typeof (tag_id) != 'string') {
     params = tag_id;
     tag_id = 'particles-js';
   }
 
   /* no id? set the id to default id */
-  if(!tag_id){
+  if (!tag_id) {
     tag_id = 'particles-js';
   }
 
   /* pJS elements */
   var pJS_tag = document.getElementById(tag_id),
-      pJS_canvas_class = 'particles-js-canvas-el',
-      exist_canvas = pJS_tag.getElementsByClassName(pJS_canvas_class);
+    pJS_canvas_class = 'particles-js-canvas-el',
+    exist_canvas = pJS_tag.getElementsByClassName(pJS_canvas_class);
 
   /* remove canvas if exists into the pJS target tag */
-  if(exist_canvas.length){
-    while(exist_canvas.length > 0){
+  if (exist_canvas.length) {
+    while (exist_canvas.length > 0) {
       pJS_tag.removeChild(exist_canvas[0]);
     }
   }
@@ -112,7 +112,7 @@ window.particlesJS = function(tag_id, params){
   var canvas = document.getElementById(tag_id).appendChild(canvas_el);
 
   /* launch particle.js */
-  if(canvas != null){
+  if (canvas != null) {
     let pjs = new pJS(tag_id, params);
     let found = false;
 
@@ -132,19 +132,19 @@ window.particlesJS = function(tag_id, params){
 
 };
 
-window.particlesJS.load = function(tag_id, path_config_json, callback){
+window.particlesJS.load = function (tag_id, path_config_json, callback) {
 
   /* load json config */
   var xhr = new XMLHttpRequest();
   xhr.open('GET', path_config_json);
   xhr.onreadystatechange = function (data) {
-    if(xhr.readyState == 4){
-      if(xhr.status == 200){
+    if (xhr.readyState == 4) {
+      if (xhr.status == 200) {
         var params = JSON.parse(data.currentTarget.response);
         window.particlesJS(tag_id, params);
-        if(callback) callback();
-      }else{
-        console.log('Error pJS - XMLHttpRequest status: '+xhr.status);
+        if (callback) callback();
+      } else {
+        console.log('Error pJS - XMLHttpRequest status: ' + xhr.status);
         console.log('Error pJS - File config not found');
       }
     }
